requirements.txt
torch>=2.0
opencv-python>=4.8
numpy>=1.24
scikit-image>=0.21
tqdm>=4.65
matplotlib>=3.7
pandas>=2.0
pyyaml>=6.0
